requirements.txt
domino-py[cli]
pytest==7.4.3
pytest-cov==4.1.0